pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coflect"
version = "0.1.0"
description = "Coflect: human-in-the-loop training framework (Torch-first, multi-backend roadmap)."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Coflect Contributors" }
]
keywords = ["machine-learning", "human-in-the-loop", "xai", "pytorch", "training"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "requests>=2.31,<3",
  "numpy>=1.26,<3",
  "pillow>=10,<12",
  "torch>=2.4,<2.9",
  "torchvision>=0.19,<0.24",
]

[project.optional-dependencies]
server = [
  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.30,<1",
  "pydantic>=2.7,<3",
]
tensorflow = [
  "tensorflow>=2.17,<2.21",
  "keras>=3.0,<4",
]
jax = [
  "jax>=0.4.30,<0.8",
  "jaxlib>=0.4.30,<0.8",
]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.8,<1",
  "mypy>=1.13,<2",
  "build>=1.2,<2",
  "twine>=6,<7",
]

[project.urls]
Homepage = "https://github.com/coflect/coflect"
Documentation = "https://github.com/coflect/coflect"
Repository = "https://github.com/coflect/coflect"
Issues = "https://github.com/coflect/coflect/issues"

[project.scripts]
coflect-hitl-backend = "coflect.cli:run_hitl_backend"
coflect-hitl-trainer-torch = "coflect.cli:run_hitl_trainer_torch"
coflect-hitl-xai-worker-torch = "coflect.cli:run_hitl_xai_worker_torch"
coflect-hitl-trainer-tf = "coflect.cli:run_hitl_trainer_tf"
coflect-hitl-xai-worker-tf = "coflect.cli:run_hitl_xai_worker_tf"
coflect-hitl-forecast-worker = "coflect.cli:run_hitl_forecast_worker"
coflect-hitl-run = "coflect.cli:run_hitl_stack"

[tool.setuptools.packages.find]
where = ["."]
include = ["coflect*"]

[tool.setuptools.package-data]
"coflect.modules.hitl.backend" = ["static/*"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
strict_equality = true
ignore_missing_imports = true
